feat(study): 특정 학습대상 조회 API 구현
